Reorganize (again) tests for repo::on_disk.
Avoid deep nesting of test files when they'll be getting big.
