+++ src/repo/on_disk/tests/mod.rs
@@ -1,0 +1,1 @@
+mod new;
